wip restructure of parsing & type checking

+++ src/llvm/llvm.h
@@ -186,7 +186,7 @@ typedef struct {
   CoLLVMOS           os_type;
   CoLLVMEnvironment  env_type;
   CoLLVMObjectFormat obj_format;
-  u8                 ptr_size; // in bytes, e.g. 8 for i64
+  u32                ptr_size; // in bytes, e.g. 8 for i64
   bool               is_little_endian;
 } CoLLVMTargetInfo;
